+++ hrk_nasa_apis/lib/src/services/ssd_cneos/sbdb_cad_api/data/query_parameters/close_approach_body.dart
@@ -19,6 +19,7 @@ enum CloseApproachBody {
   neptune,
 
   /// Mentioned in documentation, but returns 400 if used
+  /// https://www.postman.com/hrishikesh-kadam/workspace/nasa-open-apis/request/2540023-bd4989e4-8fbc-4a85-b33f-7cd6c829a029
   @JsonValue('Pluto')
   pluto,
   @JsonValue('Moon')